More folder refactorization
